Remove depracated demo element watcher API, clean up methods, expose `render`
Removes the element property change watchers that observe updates to the properties and try to set the new values on the demo element. We will now rely on the demo that imports this component to do that itself (see `px-demo` component, which already does so).

Removes all the internal methods and code paths that managed updating properties. This simplifies the component and makes it easier to test/reckon about.

Exposes a public `render()` method that can be explicitly called from the outside to rebuild the demo snippet's code string on demand.